+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - [\#1029](https://github.com/cosmos/evm/pull/1029) Mark EvmCoinInfo.Decimals field as deprecated
 - [\#1079](https://github.com/cosmos/evm/pull/1079) Access Control List is now case-insensitive
 - [\#1108](https://github.com/cosmos/evm/pull/1108) Add json-rpc http server request body limit.
+- [\#1118](https://github.com/cosmos/evm/pull/1118) Cache chain denom for evm mempool
 
 ### FEATURES